Menu working
The menu seems to be correctly starting and exiting the game. Bricks now spawn on a BoardManager, which requires its own Instance. Ball is no longer launching from the paddle when loading the scene, however. It seems to be something to do with m_Started.
