+++ Assets/Scripts/BoardManager.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 036c34898e0a3314495a61a8131245d8
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
