+++ Assets/Scripts/DeathZone.cs
@@ -5,11 +5,11 @@ using UnityEngine;
 
 public class DeathZone : MonoBehaviour
 {
-    public MainManager Manager;
+    public GameObject BoardManager;
 
     private void OnCollisionEnter(Collision other)
     {
         Destroy(other.gameObject);
-        Manager.GameOver();
+        BoardManager.GetComponent<BoardManager>().GameOver();
     }
 }
