+++ Assets/Scripts/MainManager.cs
@@ -7,62 +7,30 @@ using System.IO;
 
 public class MainManager : MonoBehaviour
 {
-    public Brick BrickPrefab;
-    public int LineCount = 6;
-    public Rigidbody Ball;
-
-    public Text ScoreText;
-    public GameObject GameOverText;
     
-    private bool m_Started = false;
-    private int m_Points;
-    
-    private bool m_GameOver = false;
-
     public static MainManager Instance;
     public string PlayerName;
-    //public ScoreList highScores;
+    //public ScoreList highScores
 
     private void Awake(){
         //create a singleton MainManager
 
-        // if there is already another MainManager, destroy this one
         if(Instance != null){
             Destroy(gameObject);
             return;
         }
-
-        //otherwise, create a MainManager instance;
         Instance = this;
-        /* this.BrickPrefab = BrickPrefab;
-        this.LineCount = LineCount;
-        this.Ball = Ball; */
-
-
         DontDestroyOnLoad(gameObject);
 
-     
     }
+
+  
 
 
     // Start is called before the first frame update
     void Start()
     {
-        const float step = 0.6f;
-        int perLine = Mathf.FloorToInt(4.0f / step);
         
-        int[] pointCountArray = new [] {1,1,2,2,5,5};
-        for (int i = 0; i < LineCount; ++i)
-        {
-            for (int x = 0; x < perLine; ++x)
-            {
-                Vector3 position = new Vector3(-1.5f + step * x, 2.5f + i * 0.3f, 0);
-                var brick = Instantiate(BrickPrefab, position, Quaternion.identity);
-                brick.PointValue = pointCountArray[i];
-                brick.onDestroyed.AddListener(AddPoint);
-                brick.transform.SetParent(MainManager.Instance.transform);
-            }
-        }
     }
 
            
@@ -70,39 +38,9 @@ public class MainManager : MonoBehaviour
 
     private void Update()
     {
-        if (!m_Started)
-        {
-            if (Input.GetKeyDown(KeyCode.Space))
-            {
-                m_Started = true;
-                float randomDirection = Random.Range(-1.0f, 1.0f);
-                Vector3 forceDir = new Vector3(randomDirection, 1, 0);
-                forceDir.Normalize();
-
-                Ball.transform.SetParent(null);
-                Ball.AddForce(forceDir * 2.0f, ForceMode.VelocityChange);
-            }
-        }
-        else if (m_GameOver)
-        {
-            if (Input.GetKeyDown(KeyCode.Space))
-            {
-                SceneManager.LoadScene(SceneManager.GetActiveScene().buildIndex);
-            }
-        }
+       
     }
 
-    void AddPoint(int point)
-    {
-        m_Points += point;
-        ScoreText.text = $"Score : {m_Points}";
-    }
-
-    public void GameOver()
-    {
-        m_GameOver = true;
-        GameOverText.SetActive(true);
-    }
 
     
 }
